refactor: interface → domain 依存を許可し、表示変換をインターフェース層に移動する
check-layer-deps.sh の interface:domain 禁止を解除し、インターフェース層が
PrState を DisplayConfig に直接マッピングできるようにする。
アプリケーション層は fetch ユースケース(Result<(PrState, bool), ErrorToken>)を
提供し、インターフェース層がそれを受け取って表示文字列に変換する責務を持つ。
CapturingPresenter パターンと ErrorPresenter trait を廃止し、errors::into_token で
Result ベースのエラー処理に統一する。
